Make multiplyByInt predict result size
The upstream code always predicts there will be a carry, and
reallocates the result when the carry is zero. Assuming bit-lengths
of operands mod 32 are uniformaly distributed, that will be true
about half the time. In real life, the short operand is likely to be
very short, so it will be true less than half the time.
When it's false, we allocate about twice as much memory as a result.

Instead use the leading digit to predict a zero carry with high
confidence.

Improves performance by about 10% for the version of the linaro
pidigits benchmark I was using, and should similarly help with other
multiplications where one operand is a single digit. Pidigits does
mostly <huge>x<single digit> multiplications, but I think this should
also help for <single digit>x<single digit> BigInteger
multiplications.

May worsen performance on certain data, but that seems very unlikely
for real data with varying bit-length.

(Note the bug below was recycled. This partially addresses the 2021
issue.)

Bug: 140780742
Test: Ran benchmark version with and without CL.
Test: Ran benchmark with assertions turned into expicit throws.
Change-Id: Idbb2e4cf788b6c48b34dd85cd8bc4bb78fa862ae